requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
cloudinary==1.37.0
Django==4.2.7
django-cloudinary-storage==0.3.0
django-environ==0.11.2
django-widget-tweaks==1.5.0
idna==3.6
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
